fix function references, package names, etc. after move
